+++ juhee/src/stackandqueue/Q02.java
@@ -1,35 +1,25 @@
-package stack;
+package stackandqueue;
 
 import java.io.*;
 import java.util.Stack;
 
-/**
- * 쇠막대기
- */
-public class Q05 {
+public class Q02 {
     public static void main(String[] args) throws IOException {
         BufferedReader stdIn = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter stdOut = new BufferedWriter(new OutputStreamWriter(System.out));
-        char[] input = stdIn.readLine().toCharArray();
         Stack<Character> stack = new Stack<>();
-        stack.push(input[0]);
-        int cnt = 0;
-        for (int i = 1; i < input.length; i++) {
+        char[] input = stdIn.readLine().toCharArray();
+        for (int i = 0; i < input.length; i++) {
             if (input[i] == ')') {
-                if (input[i - 1] == input[i]) {
+                while (!stack.isEmpty() && stack.peek() != '(') {
                     stack.pop();
-                    cnt++;
-                } else {
-                    stack.pop();
-                    cnt += stack.size();
                 }
+                stack.pop();
             } else {
                 stack.push(input[i]);
             }
         }
-        stdOut.write(String.valueOf(cnt));
+        for(int i = 0; i < stack.size(); i++) stdOut.write(stack.elementAt(i));
         stdOut.flush();
-        stdOut.close();
-        stdIn.close();
     }
 }
